src/Dapr-azure-queue/python-queue-reader/requirements.txt
asyncio
azure.storage.queue